Add anim-cycle overrides
Add anim-cycle for the asset and for the layout with time-offsets randomly distributed.
Reorganize the files in the repo
